+++ src/auth.rs
@@ -24,15 +24,21 @@ pub trait Authorised {}
 impl Authorised for AuthCodeGrantPKCEFlow {}
 impl Authorised for AuthCodeGrantFlow {}
 
+/// A Spotify token.
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct Token {
+    /// The token used for authenticating every single request.
     pub access_token: AccessToken,
+    /// The token used for requesting a new access token when the current one expires.
     pub refresh_token: Option<RefreshToken>,
+    /// How long until the current token expires, in seconds.
     pub expires_in: u64,
 
     #[serde(skip)]
+    /// The UTC date and time when the token was created.
     pub created_at: DateTime<Utc>,
     #[serde(skip)]
+    /// The UTC date and time when the token will expire.
     pub expires_at: DateTime<Utc>,
 
     #[serde(deserialize_with = "oauth2::helpers::deserialize_untagged_enum_case_insensitive")]
@@ -48,23 +54,36 @@ pub struct Token {
 #[derive(Clone, Copy, Debug)]
 pub struct UnAuthenticated;
 
+/// Authorisation code flow with PKCE is the recommended authorization flow
+/// if you’re implementing authorization in any other type of application
+/// where the client secret can’t be safely stored.
 #[derive(Clone, Debug)]
 pub struct AuthCodeGrantPKCEFlow {
     pub client_id: String,
 }
 #[derive(Clone, Debug)]
 
+/// The authorisation code flow is suitable for long-running applications
+/// (e.g. web and mobile apps) where the user grants permission only once.
+///
+/// If you’re using the authorisation code flow in any type of application
+/// where the client secret can't be safely stored, then you should use the PKCE extension.
 pub struct AuthCodeGrantFlow {
     pub client_id: String,
     pub client_secret: String,
 }
 #[derive(Clone, Debug)]
 
+/// The Client Credentials flow is used in server-to-server authentication.
+///
+/// Since this flow does not include authorisation, only endpoints that
+/// do not access user information can be accessed.
 pub struct ClientCredsGrantFlow {
     pub client_id: String,
     pub client_secret: String,
 }
 
+/// Holds the URL, CSRF token and PKCE verifier.
 #[derive(Debug)]
 pub struct AuthorisationPKCE {
     pub url: Url,
@@ -72,12 +91,14 @@ pub struct AuthorisationPKCE {
     pub(crate) pkce_verifier: PkceCodeVerifier,
 }
 
+/// Holds the URL, and CSRF token..
 #[derive(Debug)]
 pub struct Authorisation {
     pub url: Url,
     pub(crate) csrf_token: CsrfToken,
 }
 
+/// A Spotify [scope](https://developer.spotify.com/documentation/web-api/concepts/scopes).
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct Scope(pub(crate) oauth2::Scope);
 
@@ -94,10 +115,12 @@ impl Token {
         }
     }
 
+    /// Returns `true` if the access token has expired.
     pub fn is_expired(&self) -> bool {
         self.created_at >= self.expires_at
     }
 
+    /// Returns `true` if a refresh token is present.
     pub fn is_refreshable(&self) -> bool {
         self.refresh_token.is_some()
     }