+++ src/endpoint/category.rs
@@ -2,11 +2,11 @@ use serde::Serialize;
 
 use crate::{
     auth::AuthFlow,
+    error::Result,
     model::{
         category::{Categories, Category},
         Page,
     },
-    Result,
 };
 
 use super::{Builder, Endpoint, Limit};
@@ -23,16 +23,19 @@ pub struct BrowseCategoryEndpoint {
 }
 
 impl<F: AuthFlow> Builder<'_, F, BrowseCategoryEndpoint> {
+    #[doc = include_str!("../docs/country.md")]
     pub fn country(mut self, country: &str) -> Self {
         self.endpoint.country = Some(country.to_owned());
         self
     }
 
+    #[doc = include_str!("../docs/locale.md")]
     pub fn locale(mut self, locale: &str) -> Self {
         self.endpoint.locale = Some(locale.to_owned());
         self
     }
 
+    #[doc = include_str!("../docs/send.md")]
     pub async fn get(self) -> Result<Category> {
         self.spotify
             .get(
@@ -52,26 +55,31 @@ pub struct BrowseCategoriesEndpoint {
 }
 
 impl<F: AuthFlow> Builder<'_, F, BrowseCategoriesEndpoint> {
+    #[doc = include_str!("../docs/country.md")]
     pub fn country(mut self, country: &str) -> Self {
         self.endpoint.country = Some(country.to_owned());
         self
     }
 
+    #[doc = include_str!("../docs/locale.md")]
     pub fn locale(mut self, locale: &str) -> Self {
         self.endpoint.locale = Some(locale.to_owned());
         self
     }
 
+    #[doc = include_str!("../docs/limit.md")]
     pub fn limit(mut self, limit: u32) -> Self {
         self.endpoint.limit = Some(Limit::new(limit));
         self
     }
 
+    #[doc = include_str!("../docs/offset.md")]
     pub fn offset(mut self, offset: u32) -> Self {
         self.endpoint.offset = Some(offset);
         self
     }
 
+    #[doc = include_str!("../docs/send.md")]
     pub async fn get(self) -> Result<Page<Category>> {
         self.spotify
             .get("/browse/categories".to_owned(), self.endpoint)