+++ src/endpoint/player.rs
@@ -1,3 +1,5 @@
+use std::marker::PhantomData;
+
 use reqwest::Method;
 use serde::Serialize;
 use serde_json::{json, Value};
@@ -5,8 +7,9 @@ use serde_json::{json, Value};
 use crate::{
     auth::AuthFlow,
     client::Body,
+    error::Result,
     model::{player::PlayHistory, CursorPage},
-    Nil, Result,
+    Nil,
 };
 
 use super::{Builder, Endpoint};
@@ -17,7 +20,7 @@ impl Endpoint for SeekToPositionEndpoint {}
 impl Endpoint for SetRepeatModeEndpoint {}
 impl Endpoint for SetPlaybackVolumeEndpoint {}
 impl Endpoint for ToggleShuffleEndpoint {}
-impl Endpoint for RecentlyPlayedTracksEndpoint {}
+impl<T: TimestampMarker> Endpoint for RecentlyPlayedTracksEndpoint<T> {}
 impl Endpoint for AddItemToQueueEndpoint {}
 
 #[derive(Clone, Copy, Debug, Default, Serialize)]
@@ -29,6 +32,30 @@ pub enum RepeatMode {
     Off,
 }
 
+mod private {
+    use super::{After, Before, Unspecified};
+
+    pub trait Sealed {}
+
+    impl Sealed for After {}
+    impl Sealed for Before {}
+    impl Sealed for Unspecified {}
+}
+
+pub trait TimestampMarker: private::Sealed {}
+impl TimestampMarker for Before {}
+impl TimestampMarker for After {}
+impl TimestampMarker for Unspecified {}
+
+#[derive(Clone, Copy, Debug, Default)]
+pub struct After;
+
+#[derive(Clone, Copy, Debug, Default)]
+pub struct Before;
+
+#[derive(Clone, Copy, Debug, Default)]
+pub struct Unspecified;
+
 #[derive(Clone, Debug, Default, Serialize)]
 pub struct TransferPlaybackEndpoint {
     pub(crate) device_ids: Vec<String>,
@@ -36,12 +63,15 @@ pub struct TransferPlaybackEndpoint {
 }
 
 impl<F: AuthFlow> Builder<'_, F, TransferPlaybackEndpoint> {
+    /// If `true`, ensure playback happens on the new device.
+    /// Otherwise, keep the current playback state.
     pub fn play(mut self, play: bool) -> Self {
         self.endpoint.play = Some(play);
         self
     }
 
-    pub async fn transfer(self) -> Result<Nil> {
+    #[doc = include_str!("../docs/send.md")]
+    pub async fn send(self) -> Result<Nil> {
         self.spotify
             .put("/me/player".to_owned(), Body::Json(self.endpoint))
             .await
@@ -59,32 +89,38 @@ pub struct StartPlaybackEndpoint {
 }
 
 impl<F: AuthFlow> Builder<'_, F, StartPlaybackEndpoint> {
+    #[doc = include_str!("../docs/device_id.md")]
     pub fn device_id(mut self, device_id: &str) -> Self {
         self.endpoint.device_id = Some(format!("?device_id={device_id}"));
         self
     }
 
+    /// The *URI* of the context to play. Valid contexts are albums, artists and playlists.
     pub fn context_uri(mut self, context_uri: &str) -> Self {
         self.endpoint.context_uri = Some(context_uri.to_owned());
         self
     }
 
+    /// The *URI*s of the tracks to play.
     pub fn uris(mut self, uris: &[&str]) -> Self {
         self.endpoint.uris = Some(uris.iter().map(ToString::to_string).collect());
         self
     }
 
+    #[doc = include_str!("../docs/offset.md")]
     pub fn offset(mut self, offset: u32) -> Self {
         self.endpoint.offset = Some(json!({ "position": offset }));
         self
     }
 
+    /// The position at which to start/resume the playback.
     pub fn position_ms(mut self, position_ms: u32) -> Self {
         self.endpoint.position_ms = Some(position_ms);
         self
     }
 
-    pub async fn start(self) -> Result<Nil> {
+    #[doc = include_str!("../docs/send.md")]
+    pub async fn send(self) -> Result<Nil> {
         let device_id = self.endpoint.device_id.as_deref().unwrap_or("");
 
         self.spotify
@@ -103,12 +139,14 @@ pub struct SeekToPositionEndpoint {
 }
 
 impl<F: AuthFlow> Builder<'_, F, SeekToPositionEndpoint> {
+    #[doc = include_str!("../docs/device_id.md")]
     pub fn device_id(mut self, device_id: &str) -> Self {
         self.endpoint.device_id = Some(device_id.to_owned());
         self
     }
 
-    pub async fn seek(self) -> Result<Nil> {
+    #[doc = include_str!("../docs/send.md")]
+    pub async fn send(self) -> Result<Nil> {
         self.spotify
             .request(
                 Method::PUT,
@@ -127,12 +165,14 @@ pub struct SetRepeatModeEndpoint {
 }
 
 impl<F: AuthFlow> Builder<'_, F, SetRepeatModeEndpoint> {
+    #[doc = include_str!("../docs/device_id.md")]
     pub fn device_id(mut self, device_id: &str) -> Self {
         self.endpoint.device_id = Some(device_id.to_owned());
         self
     }
 
-    pub async fn set(self) -> Result<Nil> {
+    #[doc = include_str!("../docs/send.md")]
+    pub async fn send(self) -> Result<Nil> {
         self.spotify
             .request(
                 Method::PUT,
@@ -151,12 +191,14 @@ pub struct SetPlaybackVolumeEndpoint {
 }
 
 impl<F: AuthFlow> Builder<'_, F, SetPlaybackVolumeEndpoint> {
+    #[doc = include_str!("../docs/device_id.md")]
     pub fn device_id(mut self, device_id: &str) -> Self {
         self.endpoint.device_id = Some(device_id.to_owned());
         self
     }
 
-    pub async fn set(self) -> Result<Nil> {
+    #[doc = include_str!("../docs/send.md")]
+    pub async fn send(self) -> Result<Nil> {
         self.spotify
             .request(
                 Method::PUT,
@@ -175,12 +217,14 @@ pub struct ToggleShuffleEndpoint {
 }
 
 impl<F: AuthFlow> Builder<'_, F, ToggleShuffleEndpoint> {
+    #[doc = include_str!("../docs/device_id.md")]
     pub fn device_id(mut self, device_id: &str) -> Self {
         self.endpoint.device_id = Some(device_id.to_owned());
         self
     }
 
-    pub async fn toggle(self) -> Result<Nil> {
+    #[doc = include_str!("../docs/send.md")]
+    pub async fn send(self) -> Result<Nil> {
         self.spotify
             .request(
                 Method::PUT,
@@ -193,28 +237,49 @@ impl<F: AuthFlow> Builder<'_, F, ToggleShuffleEndpoint> {
 }
 
 #[derive(Clone, Debug, Default, Serialize)]
-pub struct RecentlyPlayedTracksEndpoint {
+pub struct RecentlyPlayedTracksEndpoint<T: TimestampMarker = Unspecified> {
     pub(crate) limit: Option<u32>,
     pub(crate) after: Option<u64>,
     pub(crate) before: Option<u64>,
+    marker: PhantomData<T>,
 }
 
-impl<F: AuthFlow> Builder<'_, F, RecentlyPlayedTracksEndpoint> {
+impl<'a, F: AuthFlow> Builder<'a, F, RecentlyPlayedTracksEndpoint<Unspecified>> {
+    /// A Unix timestamp in miliseconds. Returns all items after (but not including) this cursor position.
+    pub fn after(self, after: u64) -> Builder<'a, F, RecentlyPlayedTracksEndpoint<After>> {
+        Builder {
+            spotify: self.spotify,
+            endpoint: RecentlyPlayedTracksEndpoint {
+                limit: self.endpoint.limit,
+                after: Some(after),
+                before: self.endpoint.before,
+                marker: PhantomData,
+            },
+        }
+    }
+
+    /// A Unix timestamp in miliseconds. Returns all items before (but not including) this cursor position.
+    pub fn before(self, before: u64) -> Builder<'a, F, RecentlyPlayedTracksEndpoint<Before>> {
+        Builder {
+            spotify: self.spotify,
+            endpoint: RecentlyPlayedTracksEndpoint {
+                limit: self.endpoint.limit,
+                after: self.endpoint.after,
+                before: Some(before),
+                marker: PhantomData,
+            },
+        }
+    }
+}
+
+impl<F: AuthFlow, T: TimestampMarker> Builder<'_, F, RecentlyPlayedTracksEndpoint<T>> {
+    #[doc = include_str!("../docs/limit.md")]
     pub fn limit(mut self, limit: u32) -> Self {
         self.endpoint.limit = Some(limit);
         self
     }
 
-    pub fn after(mut self, after: u64) -> Self {
-        self.endpoint.after = Some(after);
-        self
-    }
-
-    pub fn before(mut self, before: u64) -> Self {
-        self.endpoint.before = Some(before);
-        self
-    }
-
+    #[doc = include_str!("../docs/send.md")]
     pub async fn get(self) -> Result<CursorPage<PlayHistory>> {
         self.spotify
             .get("/me/player/recently-played".to_owned(), self.endpoint)
@@ -229,12 +294,14 @@ pub struct AddItemToQueueEndpoint {
 }
 
 impl<F: AuthFlow> Builder<'_, F, AddItemToQueueEndpoint> {
+    #[doc = include_str!("../docs/device_id.md")]
     pub fn device_id(mut self, device_id: &str) -> Self {
         self.endpoint.device_id = Some(device_id.to_owned());
         self
     }
 
-    pub async fn add(self) -> Result<Nil> {
+    #[doc = include_str!("../docs/send.md")]
+    pub async fn send(self) -> Result<Nil> {
         self.spotify
             .request(
                 Method::POST,