+++ src/endpoint/search.rs
@@ -1,6 +1,6 @@
 use serde::Serialize;
 
-use crate::{auth::AuthFlow, model::search::SearchResults, Result};
+use crate::{auth::AuthFlow, error::Result, model::search::SearchResults};
 
 use super::{Builder, Endpoint, Limit};
 
@@ -22,26 +22,34 @@ pub struct SearchEndpoint {
 }
 
 impl<F: AuthFlow> Builder<'_, F, SearchEndpoint> {
+    #[doc = include_str!("../docs/market.md")]
     pub fn market(mut self, market: &str) -> Self {
         self.endpoint.market = Some(market.to_owned());
         self
     }
 
+    #[doc = include_str!("../docs/limit.md")]
     pub fn limit(mut self, limit: u32) -> Self {
         self.endpoint.limit = Some(Limit::new(limit));
         self
     }
 
+    #[doc = include_str!("../docs/offset.md")]
     pub fn offset(mut self, offset: u32) -> Self {
         self.endpoint.offset = Some(offset);
         self
     }
 
+    /// If specified, it signals that the client can play externally hosted audio content,
+    /// and marks the content as playable in the response.
+    ///
+    /// By default externally hosted audio content is marked as unplayable in the response.
     pub fn include_external(mut self, include_external: bool) -> Self {
         self.endpoint.include_external = Some(include_external);
         self
     }
 
+    #[doc = include_str!("../docs/send.md")]
     pub async fn get(self) -> Result<SearchResults> {
         self.spotify.get("/search".to_owned(), self.endpoint).await
     }